requirements.txt
discord.py
aiomysql
yt-dlp
spotipy
easy-pil
python-dotenv
aiohttp
